+++ Sources/SuperwallKit/Analytics/Superwall Placement/SuperwallEvent.swift
@@ -229,8 +229,8 @@ public enum SuperwallEvent {
   /// When a response from the network fails to decode.
   case networkDecodingFail
 
-  /// When the attribution properties are set.
-  case attributionProps(_ attributionProps: [String: Any])
+  /// When the integration attributes are set.
+  case integrationAttributes(_ attributes: [String: Any])
 
   var canImplicitlyTriggerPaywall: Bool {
     switch self {
@@ -397,8 +397,8 @@ extension SuperwallEvent {
       return .init(objcEvent: .enrichmentComplete)
     case .networkDecodingFail:
       return .init(objcEvent: .networkDecodingFail)
-    case .attributionProps:
-      return .init(objcEvent: .attributionProps)
+    case .integrationAttributes:
+      return .init(objcEvent: .integrationAttributes)
     }
   }
 }