+++ Sources/SuperwallKit/Analytics/Superwall Placement/SuperwallEventObjc.swift
@@ -208,8 +208,8 @@ public enum SuperwallEventObjc: Int, CaseIterable {
   /// When a response from the network fails to decode.
   case networkDecodingFail
 
-  /// When the attribution properties are set.
-  case attributionProps
+  /// When the integration attributes are set.
+  case integrationAttributes
 
   public init(event: SuperwallEvent) {
     self = event.backingData.objcEvent
@@ -341,8 +341,8 @@ public enum SuperwallEventObjc: Int, CaseIterable {
       return "enrichment_complete"
     case .networkDecodingFail:
       return "networkDecoding_fail"
-    case .attributionProps:
-      return "attribution_props"
+    case .integrationAttributes:
+      return "integration_attributes"
     }
   }
 }